chore(client): move client creation function and tests to same context
